+++ example/hello-world.php
@@ -1,5 +1,3 @@
 <?php
 
-sleep(5);
-
 echo "Hello, World!\n";