+++ Changelog.md
@@ -1,3 +1,11 @@
 # Changelog
 
 ## 0.0.1 - release
+
+### Feature
+ - support standard expression for job
+ - improve time matching logic in parser options
+ - add file driver for parser
+ - add parser
+
+### Fixes